feat: move components outside api-traffic-v4
