+++ gravitee-apim-console-webui/src/management/api/api-audit-logs/api-audit-logs.component.html
