+++ gravitee-apim-console-webui/src/management/api/api-audit-logs/components/api-audits-table/api-audits-table.component.html
