+++ gravitee-apim-console-webui/src/management/api/api-audit-logs/components/api-events-table/api-events-table.component.html
