+++ gravitee-apim-console-webui/src/management/api/api-audit-logs/components/api-events-table/api-events-table.component.scss
@@ -2,7 +2,7 @@
 @use '@angular/material' as mat;
 @use '@gravitee/ui-particles-angular' as gio;
 
-@use '../../../../../../scss/gio-layout' as gio-layout;
+@use '../../../../../scss/gio-layout' as gio-layout;
 $foreground: map.get(gio.$mat-legacy-theme, foreground);
 
 :host {
