+++ gravitee-apim-console-webui/src/management/api/api-audit-logs/components/api-events-table/api-events-table.component.ts
@@ -13,16 +13,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { CommonModule } from '@angular/common';
-import { NgModule } from '@angular/core';
-import { MatCardModule } from '@angular/material/card';
-import { GioLoaderModule } from '@gravitee/ui-particles-angular';
+import { Component, Input } from '@angular/core';
 
-import { AuditLogsComponent } from './audit-logs.component';
-import { ApiAuditsFilterFormModule, ApiAuditsTableModule, ApiEventsTableModule } from './components';
+import { Event } from '../../../../../entities/management-api-v2/event';
 
-@NgModule({
-  declarations: [AuditLogsComponent],
-  imports: [CommonModule, MatCardModule, GioLoaderModule, ApiAuditsFilterFormModule, ApiAuditsTableModule, ApiEventsTableModule],
+@Component({
+  selector: 'api-events-table',
+  templateUrl: './api-events-table.component.html',
+  styleUrls: ['./api-events-table.component.scss'],
 })
-export class AuditLogsModule {}
+export class ApiEventsTableComponent {
+  protected displayedColumns = ['icon', 'type', 'createdAt', 'user'];
+  private _events?: Event[];
+
+  public get events(): Event[] | undefined {
+    return this._events;
+  }
+  @Input()
+  public set events(value: Event[] | undefined) {
+    this._events = value;
+  }
+}
