+++ gravitee-apim-console-webui/src/management/api/api-traffic-v4/api-runtime-logs-v4.module.ts
@@ -19,9 +19,8 @@ import { NgModule } from '@angular/core';
 import { ApiRuntimeLogsSettingsModule } from './runtime-logs-settings/api-runtime-logs-settings.module';
 import { ApiRuntimeLogsModule } from './runtime-logs/api-runtime-logs.module';
 import { ApiRuntimeLogsDetailsModule } from './runtime-logs-details/api-runtime-logs-details.module';
-import { AuditLogsModule } from './audit-logs/audit-logs.module';
 
 @NgModule({
-  imports: [CommonModule, ApiRuntimeLogsModule, ApiRuntimeLogsSettingsModule, ApiRuntimeLogsDetailsModule, AuditLogsModule],
+  imports: [CommonModule, ApiRuntimeLogsModule, ApiRuntimeLogsSettingsModule, ApiRuntimeLogsDetailsModule],
 })
 export class ApiRuntimeLogsV4Module {}
