+++ gravitee-apim-console-webui/src/management/api/apis.module.ts
@@ -47,6 +47,7 @@ import { ApiDocumentationModule } from './documentation/api-documentation.module
 import { ApiUserGroupModule } from './user-group-access/api-user-group.module';
 import { ApiAuditListModule } from './api-audit-list/api-audit-list.module';
 import { ApiDeploymentConfigurationModule } from './deployment-configuration-v4/api-deployment-configuration.module';
+import { ApiAuditLogsModule } from './api-audit-logs/api-audit-logs.module';
 
 import { SpecificJsonSchemaTypeModule } from '../../shared/components/specific-json-schema-type/specific-json-schema-type.module';
 import { DocumentationModule } from '../../components/documentation/documentation.module';
@@ -84,6 +85,7 @@ import { AlertsModule } from '../../components/alerts/alerts.module';
     ApiResourcesModule,
     ApiResponseTemplatesModule,
     ApiRuntimeLogsV4Module,
+    ApiAuditLogsModule,
     ApiSubscriptionsModule,
     ApiV4PolicyStudioModule,
     ApiUserGroupModule,
